refactored old install.py into packtet.py as a pseudo packet manager for viur tools
